csvio, clone, drivers: the tier's fields reach the file formats and the wire seam
The native CSV schema gains a second version, chosen the way the
codeplug writer chooses a schema: the version-1 header while no
tier-added field records anything, version 2 as soon as one does. An
FT-710 export is byte-identical to the pre-tier exporter's, pinned
against a golden that exporter produced. Import accepts both versions
and parses freq_hz as uint64, which it never could before.

The CHIRP importer's duplex, offset, split and tone mapping becomes
capability-aware in both branches: a radio that reaches the fields keeps
its offset, stores CHIRP's split as a transmit frequency and maps DTCS
onto a code and polarity; a radio that does not takes the original path,
refusals and wordings unchanged.

The Yaesu drivers read every new field back Unavailable and their banks
list none of them, so they report Unsupported. core/cat keeps its uint32
frequency and gains the one checked conversion at the driver seam.
clone's readback verification extends to the ten new fields under
TagDisplay's mutual-knowledge rule.

+++ app/fileio_test.go
@@ -207,7 +207,7 @@ func freqOfSlot(cp *codeplug.Codeplug, slot string) int64 {
 // matching syntheticWorking's channel shape.
 func editChannelFreq(slot string, freq int64) codeplug.Channel {
 	return codeplug.Channel{Slot: slot, Data: &codeplug.ChannelData{
-		FreqHz: uint32(freq), Mode: "USB", CTCSS: "OFF", Shift: "SIMPLEX",
+		FreqHz: uint64(freq), Mode: "USB", CTCSS: "OFF", Shift: "SIMPLEX",
 		CTCSSTone:  codeplug.ToneField{State: codeplug.Unknown},
 		TagDisplay: codeplug.BoolField{State: codeplug.Known, Value: false},
 		ScanSkip:   codeplug.BoolField{State: codeplug.Unknown},
