+++ cmd/rigprog/write_inprocess_test.go
@@ -103,7 +103,7 @@ func minimalFactoryPopulated() map[string]*codeplug.ChannelData {
 // every FieldState-carrying field Unknown (nothing inexpressible
 // requested), everything else set to plain, codec-expressible values.
 // Mirrors core/clone/helpers_test.go's helper of the same name/shape.
-func writableChannel(slot string, freqHz uint32, tag string) codeplug.Channel {
+func writableChannel(slot string, freqHz uint64, tag string) codeplug.Channel {
 	return codeplug.Channel{
 		Slot: slot,
 		Data: &codeplug.ChannelData{
